# services/pokemon-service-node/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Pokémons - Microservicio Node.js",
    "version": "1.0.0",
    "description": "Microservicio desarrollado en Node.js y Express para consultar los Pokémon almacenados en la base de datos relacional PostgreSQL en Render."
  },
  "servers": [
    {
      "url": "/",
      "description": "Servidor actual"
    }
  ],
  "tags": [
    {
      "name": "Pokémon",
      "description": "Endpoints para consultar los Pokémon de la base de datos"
    },
    {
      "name": "Estado del Servidor",
      "description": "Verificación del servicio y conexión a la base de datos"
    }
  ],
  "paths": {
    "/api/pokemon": {
      "get": {
        "tags": ["Pokémon"],
        "summary": "Listar todos los Pokémon",
        "description": "Retorna la lista completa de todos los Pokémon registrados en la base de datos PostgreSQL.",
        "responses": {
          "200": {
            "description": "Lista de Pokémon obtenida con éxito",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pokemon"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al consultar la base de datos"
          }
        }
      }
    },
    "/api/pokemon/{nameOrId}": {
      "get": {
        "tags": ["Pokémon"],
        "summary": "Buscar un Pokémon",
        "description": "Busca un Pokémon por su número de ID o por su nombre (por ejemplo: bulbasaur, charmander, squirtle).",
        "parameters": [
          {
            "name": "nameOrId",
            "in": "path",
            "required": true,
            "description": "ID numérico o nombre del Pokémon",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pokémon encontrado con éxito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pokemon"
                }
              }
            }
          },
          "404": {
            "description": "El Pokémon no existe en la base de datos"
          },
          "500": {
            "description": "Error al consultar la base de datos"
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "tags": ["Estado del Servidor"],
        "summary": "Verificar conexión",
        "description": "Indica si el servidor está en línea y si está conectado correctamente a PostgreSQL.",
        "responses": {
          "200": {
            "description": "Servidor funcionando correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "database": {
                      "type": "string",
                      "example": "connected (postgresql)"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-10-02T19:30:00.000Z"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Pokemon": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Número identificador del Pokémon",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Nombre del Pokémon",
            "example": "bulbasaur"
          },
          "height": {
            "type": "integer",
            "description": "Altura del Pokémon en decímetros",
            "example": 7
          },
          "weight": {
            "type": "integer",
            "description": "Peso del Pokémon en hectogramos",
            "example": 69
          },
          "sprites": {
            "type": "object",
            "description": "Imágenes del Pokémon",
            "properties": {
              "front_default": {
                "type": "string",
                "example": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/1.png"
              }
            }
          },
          "moves": {
            "type": "array",
            "description": "Lista de movimientos principales",
            "items": {
              "type": "object",
              "properties": {
                "move": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "razor-wind"
                    },
                    "url": {
                      "type": "string",
                      "example": "https://pokeapi.co/api/v2/move/13/"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}